tests/wpt/tests/css/css-gaps/animation/gap-decorations-outset-neutral-keyframe-002.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>gap decorations column-rule-outset neutral keyframe</title>
    <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#outset">
    <meta name="assert" content="gap decorations column-rule-outset value list supports neutral keyframe.">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>
  <body>
    <div id="target"></div>
    <script>
      test(() => {
        target.style.columnRuleStyle = 'solid';
        target.style.columnRuleOutset = '1px';
        var animation = target.animate([{}, {columnRuleOutset: '11px'}], 1000);
        animation.pause();
        animation.currentTime = 500;
        assert_equals(getComputedStyle(target).columnRuleOutset, '6px');
      }, 'gap decorations column-rule-outset value list supports neutral keyframe.');
    </script>
  </body>
</html>
